+++ CompatibilityLib/base/src/main/java/com/elmakers/mine/bukkit/utility/platform/base/CompatibilityUtilsBase.java
@@ -120,21 +120,32 @@ public abstract class CompatibilityUtilsBase implements CompatibilityUtils {
         }
     }
 
-    @Override
     public boolean applyPotionEffect(LivingEntity entity, PotionEffect effect) {
         // Avoid nerfing existing effects
         boolean applyEffect = true;
         Collection<PotionEffect> currentEffects = entity.getActivePotionEffects();
         for (PotionEffect currentEffect : currentEffects) {
-            if (currentEffect.getType().equals(effect.getType())) {
-                if (effect.getAmplifier() < 0) {
-                    applyEffect = false;
-                    break;
-                } else if (currentEffect.getAmplifier() > effect.getAmplifier() || effect.getDuration() > Integer.MAX_VALUE / 4) {
-                    applyEffect = false;
-                    break;
-                }
+            // Look for the effect type we care about
+            if (!currentEffect.getType().equals(effect.getType())) {
+                continue;
             }
+
+            int currentAmplifier = currentEffect.getAmplifier();
+            int newAmplifier = effect.getAmplifier();
+            // Don't replace negative effects, except with more negative effects
+            if (currentAmplifier < 0)  {
+                applyEffect = newAmplifier < currentAmplifier;
+            } else if (currentEffect.getDuration() > Integer.MAX_VALUE / 4 && effect.getDuration() <= Integer.MAX_VALUE / 4) {
+                // Don't replace infinite effects, except with other infinite effects
+                applyEffect = false;
+            } else if (newAmplifier > 0 && newAmplifier < currentAmplifier) {
+                // Don't replace if the new amplifier is positive but less than the current amplifier
+                applyEffect = false;
+            }
+
+            // Strictly speaking active effects come from a Map, so there should never be more
+            // than one effect of the same type, even though Bukkit exposes it as a List
+            break;
         }
         if (applyEffect) {
             entity.addPotionEffect(effect, true);